Better placement of formatting code
