+++ scanomatic/ui_server_data/js/ccc/helpers.js
@@ -21,6 +21,16 @@ export const featureColors = {
 };
 
 
+export const valueFormatter = (value, fixed = 0) => {
+    if (value === 0) {
+        return value.toFixed(0);
+    }
+    const exponent = Math.floor(Math.log10(Math.abs(value)));
+    const number = (value / (10 ** exponent)).toFixed(fixed);
+    return `${number} x 10^${exponent.toFixed(0)}`;
+};
+
+
 export function getDataUrlfromUrl(src, callback) {
     var img = new Image();
     img.crossOrigin = 'Anonymous';
